% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autostereogram.R
\name{autostereogram}
\alias{autostereogram}
\title{autostereogram}
\usage{
autostereogram(depth.map, pattern, repetitions)
}
\arguments{
\item{depth.map}{Depth map representing the 3D shape. Possible values are:  \n  Numeric matrix. Elements with large value appear closer. \n Character containing path to greyscale .bmp (24bits) picture. The whiter a pixel is, the closer it appears.}

\item{pattern}{Optional. Path to a picture representing the pattern used to create the autostereogram. If not provided, a random dot autostereogram is generated.}

\item{repetitions}{Numeric. Number of times the pattern should be repeated horizontally.}
}
\value{
array representing image
}
\description{
autostereogram
}
