% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.map.R
\name{import.map}
\alias{import.map}
\title{Internal function. Import a greyscale picture and convert it in a matrix. Supported picture formats are PNG, JPEG and BMP.}
\usage{
import.map(path)
}
\arguments{
\item{path}{Character indicating path to depth map (greyscale picture)}
}
\value{
Matrix where elements are in [0,1]. White pixel become 1, and black 0.
}
\description{
Internal function. Import a greyscale picture and convert it in a matrix. Supported picture formats are PNG, JPEG and BMP.
}
