% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autostereogram.R
\name{autostereogram}
\alias{autostereogram}
\title{Creates autostereogram based on a depth map and a pattern. If no pattern is provided, a random-dot autostereogram is created. The object returned is a cimg from the image package. Supported picture formats are PNG, JPEG and BMP.}
\usage{
autostereogram(depth.map, pattern, repetitions)
}
\arguments{
\item{depth.map}{Depth map representing the 3D shape. Possible values are:  a) Numeric matrix where lements with large value appear closer. b) Character containing path to greyscale .bmp .jpg or .png picture. The whiter a pixel is, the closer it appears.}

\item{pattern}{Optional. Path to a picture (.bmp .jpg or .png) representing the pattern used to create the autostereogram. If not provided, a random dot autostereogram is generated.}

\item{repetitions}{Optional. Number of times the pattern should be repeated horizontally. Default is round(width(depth.map)/100)}
}
\value{
cimg image
}
\description{
Creates autostereogram based on a depth map and a pattern. If no pattern is provided, a random-dot autostereogram is created. The object returned is a cimg from the image package. Supported picture formats are PNG, JPEG and BMP.
}
\examples{
# Plot bivariate gaussian distribution
x <- seq(-2.5, 2.5, 0.02)
y <- seq(-2.5, 2.5, 0.05)
depth.map <- matrix(data=NA, nrow=length(x), ncol=length(y))
for(i in 1:length(x)) {
  for(j in 1:length(y)) {?
    depth.map[i,j] <-  (1/(2*pi)) * exp( -((x[i])^2 + (y[j])^2)/2 )
  }
}
gaussian <- autostereogram(depth.map)
plot(gaussian, axes=FALSE)
}
