% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.map.R
\name{import.map}
\alias{import.map}
\title{import.map}
\usage{
import.map(path)
}
\arguments{
\item{path}{Character indicating path to depth map (BW picture)}
}
\value{
matrix where elements are in [0,1]. White pixel become 1, and black 0.
}
\description{
import.map
}
